Permutation of String
Write a program to print all permutations of a given string
A permutation, also called an “arrangement number” or “order,” is a
rearrangement of the elements of an ordered list S into a one-to-one
correspondence with S itself. A string of length n has n! permutation.
Source: Mathword(http://mathworld.wolfram.com/Permutation.html)

Below are the permutations of string ABC.
ABC, ACB, BAC, BCA, CAB, CBA

http://www.geeksforgeeks.org/write-a-c-program-to-print-all-permutations
-of-a-given-string/